organize redux files
