+++ Frontend/src/services/SearchEngine.js
@@ -2,12 +2,13 @@ import LinkConst from "./LinkConst";
 import { connect } from 'react-redux';
 import { useDispatch } from "react-redux";
 import { useEffect } from "react";
+import * as SearchActions from '../store/actions/search';
 
 const SearchAPIRequest = ({searchContent, callSearchApi}) => {      
     const dispatch = useDispatch()
     useEffect(() => {
         if(callSearchApi) {
-            dispatch(ClearOldRequests())
+            dispatch(SearchActions.ClearOldRequests())
             fetch(LinkConst.SearchContent, { 
                 method: 'POST',
                 headers: {
@@ -18,27 +19,11 @@ const SearchAPIRequest = ({searchContent, callSearchApi}) => {
             }).
             then(reponse => reponse.json()).
             then(reponse => reponse.map((info) => {
-                dispatch(setSearchData(info.img, info.link, info.title))
+                dispatch(SearchActions.setSearchData(info.img, info.link, info.title))
             }))
         }
     },[callSearchApi])
 };
- 
-function setSearchData(image, link, title) {
-    return { 
-        type : 'SET_CONTENT_SEARCH',
-        image,
-        link,
-        title
-    }
-}
-
-function ClearOldRequests() {
-    return {
-        type : 'CLEAR_OLD_REQUESTS'
-    }
-}
-
 
 export default connect()(SearchAPIRequest)
