+++ Frontend/src/store/actions/play.js
@@ -1,0 +1,9 @@
+export function setDataMusic(image, artist, title, trackData) {
+    return { 
+        type : 'SET_MUSIC_DATA',
+        image,
+        artist,
+        title,
+        trackData
+    }
+}
