+++ Frontend/src/store/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import search from './search';
+import play from './play';
+
+export default combineReducers({
+    search,
+    play,
+})
